fix: 평가용 프롬프트 수정
