--- dnd-14th-1-iOS/Domain/Model/Diagnosis/EfficiencyDiagnosis.swift
@@ -1,12 +1,0 @@
-//
-//  EfficiencyDiagnosis.swift
-//  dnd-14th-1-iOS
-//
-//  Created by a on 2/25/26.
-//
-
-import FoundationModels
-
-struct EfficiencyDiagnosis { 
-    let efficiency: EfficiencyType
-}
